Move csmc_standard to llamppl; rename csmc.py → csmc_memory.py
Vicky's PR #144 review asked for csmc_standard to live alongside
smc_standard in llamppl. Done: the inference loop is now imported from
llamppl (pinned at the relevant commit until a release lands). What
stays here is the genlm-control-flavored adapter — the
RetainedTokenSampler that knows about TokenSampler / parent.token_ctx /
parent.score and so can't move upstream. csmc.py renamed to
csmc_memory.py since the file no longer contains the C-SMC algorithm,
just the retained-token wrapper.

Also drops the EM/TRICE framing from docs/csmc.md per Vicky's first
comment on the PR.
