+++ pyproject.toml
@@ -18,7 +18,7 @@ dependencies = [
     "genlm-grammar>=0.2.0",
     "genlm-backend>=0.2.0",
     "genlm-bytes>=0.2.0",
-    "llamppl>=0.2.2",
+    "llamppl @ git+https://github.com/genlm/llamppl@9a26c29",
     "arsenal>=3.1.3",
     "IPython",
     "numpy",
